+++ tests/src/Fixtures/EnumAndStringableDto/ObjectWithStringablePropertyDto.php
@@ -13,18 +13,7 @@ declare(strict_types=1);
 
 namespace Rekalogika\Mapper\Tests\Fixtures\EnumAndStringableDto;
 
-use Rekalogika\Mapper\Tests\Fixtures\EnumAndStringable\SomeBackedEnum;
-use Rekalogika\Mapper\Tests\Fixtures\EnumAndStringable\SomeEnum;
-
-class ObjectWithEnumStringablePropertyDto
+class ObjectWithStringablePropertyDto
 {
     public ?string $stringable = null;
-
-    public ?string $backedEnum = null;
-
-    public ?string $unitEnum = null;
-
-    public ?SomeBackedEnum $stringBackedEnum = null;
-
-    public ?SomeEnum $stringUnitEnum = null;
 }